+++ src/components/material/Material.scss
@@ -50,7 +50,7 @@
 
   .document-download-toggle {
     position: absolute;
-    z-index: 5;
+    z-index: 51;
     right: 0;
     bottom: -5em;
     width: 12em;
